% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_asc_from_raster.R
\name{sdm_asc_from_raster}
\alias{sdm_asc_from_raster}
\title{Raster conversion functions}
\usage{
sdm_asc_from_raster(x)
}
\arguments{
\item{x}{is an object of class 'RasterLayer'}
}
\value{
Returns an object of class requested.
}
\description{
Raster conversion functions
}
\details{
\code{sdm_asc_from_raster} is an adaptation of
\code{asc.from.raster} from SDMTools; extracts data from objects of
class 'RasterLayer' (raster package) into an object of class 'asc'.
}
\author{
Gopi Goteti
}
